Simplifying things
One of the critcisms recieved during my code review was that having everything in main was probably not a good idea.
